requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
mysql-connector-python